Calculate the measurement matrix given N observations and G grid points

tau is a vector of calendar dates indexed from g = 1,2,...,G. phi_m is a vector of n = 1,2,...,N radiocarbon determinations (fraction modern) with associated uncertainties sig_m. Calculate the measurement matrix M, which has dimensions N x G and for which each element is M_ig = p(phi_m,i|tau_g). The total uncertainty of the measurement comes from measurement error (SIG_M, calculated using the measurement error for each measurement) and the calibration curve error (SIG_c, calculated using the uncertainty for the calibration curve at each grid point). These uncertainties (and the associated measurements) should already be "projected" to 1950 equivalents. By default the measurement matrix, M, is multiplied by dtau so that M * f is an approximation to the integral over p(t|th) * p(phi|t) using a Riemann sum with the density calculated at the points tau and the width for each point being dtau. Alternatively, the trapezoidal rule can be used (for details see bd_calc_trapez_weights). If the spacing of tau is irregular, the trapezoidal rule must be used. An error is thrown if tau is irregularly spaced and the the trapezoidal rule is not used

bd_calc_meas_matrix(tau, phi_m, sig_m, calibDf, addCalibUnc = T, useTrapez = F)

Arguments

tau
A vector of calendar dates indexed by g

phi_m
A vector of fraction moderns indexed by i

sig_m
A vector of standard deviations for phi_m by i

calibDf
Calibration curve (see bd_load_calib_curve)

addCalibUnc
(default TRUE) Whether to add calibration uncertainty

useTrapez
(default FALSE) Whether to use the trapezoidal rule for integration