Sample from a truncated exponential distribution

Make N samples from a truncated exponential distribution with the density function r0exp(r0y). The rate r0 can be positive (growth) or negative (decay). The density function is normalized to integrate to 1 on the interval ymin to ymax.

bd_sample_trunc_exp(N, r0, ymin, ymax)

Arguments

N
The number of samples to make

r0
The rate parameter

ymin
The lower limit for truncation

ymax
The upper limit for truncation

Value

A vector of N samples between ymin and ymax