Sample from a possibly truncated Gaussian mixture

N is the number of samples to draw and th specifies the Gaussian mixture with the ordering (pi_1,...,pi_K,mu_1,...,mu_K,s_1,...,s_K), where K is the number of mixture components. Optionally, the samples are drawn on the truncated interval tau_min to tau_max. Because of limitations in the package distr, the maximum number of mixture components suppored is K=4.

sample_gauss_mix(N, th, tau_min = NA, tau_max = NA)

Arguments

N
Number of samples

th
Parameterization vector for the Gaussian mixture

tau_min
(Optional) Lower bound for samples

tau_max
(Optional) Upper bound for samples

Value

N samples from the Gaussian mixture