The SHCal13 Southern Hemisphere atmospheric radiocarbon calibration curve

A tibble::tibble containing the SHCal13 Southern Hemisphere atmospheric radiocarbon calibration curve.

shcal13

Format

An object of class tbl_df (inherits from tbl, data.frame) with 5141 rows and 5 columns.

Source

http://www.radiocarbon.org/IntCal13/

Details

shcal13 is a tibble::tbl_df object with the following fields:

CAL BP --- The calibrated year before present

14C age --- The un-calibrated 14C age of the sample, in years before present

Error --- The measurement error of the 14C age, in years before present

Delta 14C --- The normalized delta 14C (Δ14C) measurement for the sample, in per mil (‰)

Sigma --- The standard deviation (σ) of the Δ14C measurement, in per mil (‰)

References

Alan G Hogg, Quan Hua, Paul G Blackwell, Caitlin E Buck, Thomas P Guilderson, Timothy J Heaton, Mu Niu, Jonathan G Palmer, Paula J Reimer, Ron W Reimer, Christian S M Turney, Susan R H Zimmerman Radiocarbon 55(4). DOI: 10.2458/azu_js_rc.55.16783